@@ -2,9 +2,15 @@
   "name": "tagtree",
   "description": "Share all your gamer tags from one place.",
   "engines": {
-    "node": ">=15.0.0"
+    "node": "^18.17.0 || ^20.3.0 || >=21.0.0"
   },
+  "engines-annotation": "Mostly required by sharp which needs a Node-API v9 compatible runtime",
   "private": true,
+  "trustedDependencies": [
+    "sharp"
+  ],
+  "trustedDependencies-annotation": "Needed for bun to allow running install scripts",
+  "type": "module",
   "scripts": {
     "build": "qwik build",
     "build-storybook": "storybook build",
@@ -24,36 +30,33 @@
     "api": "json-server-auth db.json --r routes.json"
   },
   "devDependencies": {
-    "@builder.io/qwik": "^1.2.7",
-    "@builder.io/qwik-city": "^1.2.7",
-    "@storybook/addon-essentials": "7.2.1",
-    "@storybook/addon-links": "7.2.1",
-    "@storybook/addon-styling": "^1.3.7",
-    "@storybook/blocks": "7.2.1",
-    "@storybook/builder-vite": "7.2.1",
-    "@storybook/html": "7.2.1",
-    "@storybook/html-vite": "7.2.1",
-    "@tailwindcss/forms": "^0.5.4",
-    "@types/eslint": "8.40.2",
-    "@types/node": "^20.4.8",
-    "@typescript-eslint/eslint-plugin": "5.61.0",
-    "@typescript-eslint/parser": "5.61.0",
-    "autoprefixer": "^10.4.14",
-    "eslint": "8.44.0",
-    "eslint-plugin-qwik": "^1.2.7",
-    "eslint-plugin-storybook": "^0.6.13",
+    "@builder.io/qwik": "^1.9.1",
+    "@builder.io/qwik-city": "^1.9.1",
+    "@storybook/addon-essentials": "^8.2.8",
+    "@storybook/addon-links": "^8.2.8",
+    "@storybook/blocks": "^8.2.8",
+    "@storybook/builder-vite": "^8.2.8",
+    "@storybook/html": "^8.2.8",
+    "@storybook/html-vite": "^8.2.8",
+    "@types/eslint": "8.56.10",
+    "@types/node": "20.14.11",
+    "@typescript-eslint/eslint-plugin": "7.16.1",
+    "@typescript-eslint/parser": "7.16.1",
+    "autoprefixer": "^10.4.19",
+    "eslint": "8.57.0",
+    "eslint-plugin-qwik": "^1.9.1",
+    "postcss": "^8.4.39",
+    "prettier": "3.3.3",
+    "prettier-plugin-tailwindcss": "^0.5.4",
+    "storybook": "^8.2.8",
+    "storybook-framework-qwik": "^0.4.0",
+    "tailwindcss": "^3.4.6",
+    "typescript": "5.4.5",
+    "undici": "*",
+    "vite": "5.3.5",
+    "vite-tsconfig-paths": "^4.2.1",
     "express": "^4.18.2",
-    "json-server": "^0.17.3",
-    "json-server-auth": "^2.1.0",
-    "postcss": "^8.4.27",
-    "prettier": "2.8.8",
-    "storybook": "7.2.1",
-    "storybook-framework-qwik": "0.2.3",
-    "tailwindcss": "^3.3.3",
-    "typescript": "5.1.6",
-    "undici": "5.22.1",
-    "vite": "4.4.9",
-    "vite-tsconfig-paths": "4.2.0"
+    "@tailwindcss/forms": "^0.5.9"
   },
   "overrides": {
     "json-server-auth": {
